Add test_py_specific
